+++ Editor/CesiumIonSession.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using Reinterop;
 
 namespace CesiumForUnity
@@ -5,23 +6,24 @@ namespace CesiumForUnity
     [ReinteropNativeImplementation("CesiumForUnityNative::CesiumIonSessionImpl", "CesiumIonSessionImpl.h")]
     public partial class CesiumIonSession
     {
-        private static CesiumIonSession currentSession = null!;
-
         public delegate void GUIUpdateDelegate();
 
         public static event GUIUpdateDelegate OnConnectionUpdated;
         public static event GUIUpdateDelegate OnAssetsUpdated;
         public static event GUIUpdateDelegate OnProfileUpdated;
         public static event GUIUpdateDelegate OnTokensUpdated;
+        public static event GUIUpdateDelegate OnDefaultsUpdated;
 
-        public static CesiumIonSession Ion()
+        public CesiumIonServer server
         {
-            if (currentSession == null)
-            {
-                currentSession = new CesiumIonSession();
-            }
+            get;
+            internal set;
+        }
 
-            return currentSession;
+        public CesiumIonSession(CesiumIonServer server)
+        {
+            this.server = server;
+            this.CreateImplementation();
         }
 
         public partial bool IsConnected();
@@ -37,18 +39,24 @@ namespace CesiumForUnity
         public partial bool IsTokenListLoaded();
         public partial bool IsLoadingTokenList();
 
+        public partial bool IsDefaultsLoaded();
+        public partial bool IsLoadingDefaults();
+
         public partial void Connect();
         public partial void Resume();
         public partial void Disconnect();
 
         public partial string GetProfileUsername();
         public partial string GetAuthorizeUrl();
+        public partial string GetRedirectUrl();
+        public partial List<QuickAddItem> GetQuickAddItems();
 
         public partial void Tick();
 
         public partial void RefreshProfile();
         public partial void RefreshTokens();
         public partial void RefreshAssets();
+        public partial void RefreshDefaults();
 
         public void BroadcastConnectionUpdate()
         {
@@ -79,6 +87,14 @@ namespace CesiumForUnity
             if (OnTokensUpdated != null)
             {
                 OnTokensUpdated();
+            }
+        }
+
+        public void BroadcastDefaultsUpdate()
+        {
+            if (OnDefaultsUpdated != null)
+            {
+                OnDefaultsUpdated();
             }
         }
     }