+++ native~/Runtime/src/UnityTilesetExternals.h
@@ -6,7 +6,14 @@
 
 namespace Cesium3DTilesSelection {
 class TilesetExternals;
-}
+class CreditSystem;
+} // namespace Cesium3DTilesSelection
+
+namespace CesiumAsync {
+class AsyncSystem;
+class IAssetAccessor;
+class ITaskProcessor;
+} // namespace CesiumAsync
 
 namespace DotNet::CesiumForUnity {
 class Cesium3DTileset;
@@ -14,7 +21,11 @@ class Cesium3DTileset;
 
 namespace CesiumForUnityNative {
 
+const std::shared_ptr<CesiumAsync::IAssetAccessor>& getAssetAccessor();
+const std::shared_ptr<CesiumAsync::ITaskProcessor>& getTaskProcessor();
+CesiumAsync::AsyncSystem getAsyncSystem();
+
 Cesium3DTilesSelection::TilesetExternals
 createTilesetExternals(const DotNet::CesiumForUnity::Cesium3DTileset& tileset);
 
-}
+} // namespace CesiumForUnityNative